% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{colgrp}
\alias{colgrp}
\title{Make a Single Column Grouping}
\usage{
colgrp(x, data_name, index_name = "group")
}
\arguments{
\item{x}{A tibble}

\item{data_name}{A string, the name of the new column}

\item{index_name}{A string, the name of the new column index}
}
\value{
A grouped tibble
}
\description{
Takes a tibble and groups columns together into a single data column. All
columns that are not row indices will be grouped, and the resulting column
will be named \code{data_name}.
}
