% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_grp_cols.R
\name{def_grp_cols}
\alias{def_grp_cols}
\title{Convert a flat dataframe into a col-group structure}
\usage{
def_grp_cols(x, data_name, grp_name, ...)
}
\description{
Convert a flat dataframe into a col-group structure
}
