% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{do_ts_plot}
\alias{do_ts_plot}
\title{Plot timeseries groups of grouping object.}
\usage{
do_ts_plot(gr_obj, choice, data_choice, filter = NA, main = "Timeseries",
  date_col = "dd_dt", window = NA, ..., date_axis = "\%b-\%Y")
}
\arguments{
\item{gr_obj}{Grouping object created with \code{get_groups}}

\item{choice}{Grouping choices from grouping object}

\item{data_choice}{Calculated column to be plotted (created with the function argument in \code{get_groups})}

\item{filter}{Character string that filters the groups to be plotted}

\item{main}{Title of plot}

\item{date_col}{Group that represents date for plotting}

\item{window}{Date subset (\code{xts} subsetting)}

\item{...}{Graphics arguments}

\item{date_axis}{Date format for axis}
}
\description{
Plot timeseries groups of grouping object.
}

