% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igroup.R
\name{dot_to_arg}
\alias{dot_to_arg}
\title{Parse an argument to group_by2}
\usage{
dot_to_arg(i, dots)
}
\description{
The param (dots[i]) should be either a named vector, which creates an inapplicable group,
or a symbol making a full group. All other values are errors.
}
