% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{drop_colgrp}
\alias{drop_colgrp}
\title{Remove Column Grouping}
\usage{
drop_colgrp(data, sep = "_")
}
\arguments{
\item{data}{A tibble with a column grouping}

\item{sep}{The separator character used to construct new column names}
}
\value{
A tibble without a column grouping
}
\description{
Separate the current column grouping, represented by a data frame column, into
individual columns. The name of each new column is a concatenation of the
data column name, the argument \code{sep}, and the group name.
}
