% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igroup.R
\name{expand_igrp}
\alias{expand_igrp}
\title{Expand a df with one inapplicable grouping}
\usage{
expand_igrp(x)
}
\arguments{
\item{x}{df, that can only have exactly one grouping variable}
}
\value{
A df with only the changed (formerly inapplicable) rows
}
\description{
Expand a df with one inapplicable grouping
}
