% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igroup.R
\name{expand_igrps}
\alias{expand_igrps}
\title{Add rows to capture each grouping}
\usage{
expand_igrps(x)
}
\description{
Expansion is turning a hierarchical grouping with I-values into a flat one
without I-values.
}
