% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_df}
\alias{extract_df}
\alias{drop_df}
\title{Extract or drop an observation}
\usage{
extract_df(groupr, groups)

drop_df(groupr, groups)
}
\arguments{
\item{groupr}{The groupr object to pull from}

\item{groups}{The desired groups}
}
\value{
The desired observation, typically a dataframe or \code{xts} object
}
\description{
Pull a leaf from the \code{groupr} tree. This will typically return a dataframe or an \code{xts} object.
}
\section{Functions}{
\itemize{
\item \code{drop_df}: Drop an observation
}}

