% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_oriented.R
\name{gapply}
\alias{gapply}
\title{Apply across a groupr object}
\usage{
gapply(groupr, new_functions, is_cbind = F)
}
\arguments{
\item{groupr}{The \code{groupr} object to loop through}

\item{new_functions}{The new functions to apply}

\item{is_cbind}{Whether or not to cbind to the original dataframes}
}
\description{
An S3 method for applying a function to a groupr object. This is typically a
way to simply loop through the object which simplifies complicated
\code{lapply} functions.
}
