% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_combinations}
\alias{get_combinations}
\title{Make every combination of character vector}
\usage{
get_combinations(n, char_vec)
}
\arguments{
\item{n}{Maximum rows per combination data frame}

\item{char_vec}{Character vector to be used}
}
\value{
A list of data frames with every column a unique combination of the given character vector
}
\description{
Make every combination of character vector
}
\examples{
x <- c("foo", "bar", "baz", "potato")
y <- get_combinations(3, x)
}

