% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_groups}
\alias{get_groups}
\alias{get_functions}
\title{Get the groups or functions from a \code{groupr} object}
\usage{
get_groups(groupr)

get_functions(groupr)
}
\arguments{
\item{groupr}{The \code{groupr} object in question}
}
\description{
Simple wrapper to pull group names or function definitions.
}
