% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_xts}
\alias{make_xts}
\title{Creates \code{xts} object from grouping object.}
\usage{
make_xts(grouping_obj, choice, date_col = "dd_dt")
}
\arguments{
\item{grouping_obj}{Grouping object created with get_groups}

\item{choice}{Vector of grouping objects}
}
\value{
An \code{xts} object casted with unique group values as column names
}
\description{
Creates \code{xts} object from grouping object.
}

