% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianize.R
\name{medianize}
\alias{medianize}
\title{Medianize}
\usage{
medianize(val)
}
\arguments{
\item{val}{Array Values lista}
}
\value{
c Grouped values
}
\description{
Categorizes a continous array in two groups (low-high) using median as criteria. \cr
   With odd numbers, middle point is assigned to "low" group
}
