% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{other_label}
\alias{other_label}
\title{Automatically assign "other" to low observation count grouping values}
\usage{
other_label(df, column, percentile = 0.9, custom = NULL)
}
\arguments{
\item{df}{The dataframe to be manipulated}

\item{column}{Which column to relabel}

\item{percentile}{Which percentage to cut off the data at}

\item{custom}{A custom vector of values to reassign to "other" in the dataset}
}
\value{
The dataframe with reassigned column
}
\description{
Often times, a dataset will have groups which are distributed somewhat
exponentially - almost all rows fall into a few group values, but there are
many smaller group values for the remaining observations. For example, you
may have a dataset with employee level observations and want to use "US
State" as a group, but 90\% of the observations fall into New York,
California, Texas, and perhaps 6 other states. All remaining observations are
distributed amongst the remaining 41 states, but you might prefer to lump all
of those observations into a single bucket. This functions provides a way to
reassign all those observations to "other".
}
\examples{
summary(as.factor(permits$type_desc))
permits_cleaned <- other_label(permits, "type_desc")
summary(as.factor(permits_cleaned$type_desc))
}
