% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scientists}
\alias{scientists}
\title{Scientists}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 22 rows and 5 columns.
}
\usage{
scientists
}
\description{
Scientists dataset
}
\keyword{datasets}
