% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{sep_colgrp}
\alias{sep_colgrp}
\title{Separate Columns By a Character}
\usage{
sep_colgrp(x, sep, index_name = "group")
}
\arguments{
\item{x}{A tibble}

\item{sep}{A character delimiting the two parts of the name}

\item{index_name}{A name for the new column index}
}
\value{
A grouped tibble
}
\description{
Creates a column index by interpreting each column name as a data column name
and a group name separated by \code{sep}. Only columns that are not row indices
are used. \code{sep} must occur exactly once in each column name.
}
