% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tertilize.R
\name{tertilize}
\alias{tertilize}
\title{Medianize}
\usage{
tertilize(variable, withLabels = FALSE)
}
\arguments{
\item{variable}{Array Values list}

\item{withLabels}{Boolean Check if returns range fields/low-mid-high names}
}
\value{
c Grouped values
}
\description{
Categorizes a continous array in three groups (low-mid-high) using median as criteria. \cr
   The boundary values are included to lowest group
}
