% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igroup.R
\name{ungroup.igrouped_df}
\alias{ungroup.igrouped_df}
\title{Ungroup a Tibble With Inapplicable Groups}
\usage{
\method{ungroup}{igrouped_df}(x, ...)
}
\arguments{
\item{x}{An igrouped tibble (as created by group_by2)}

\item{...}{Ignored. All variables are removed from the grouping.}
}
\value{
A tibble with no groups. The "groups" attribute will be set to
contain one column, .rows, with a single value that lists all rows.
}
\description{
Ungroup method for tibbles that have inapplicable groups.
}
