% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_grouping_files}
\alias{write_grouping_files}
\title{Write grouping object to csv files.}
\usage{
write_grouping_files(grouping_obj, path)
}
\arguments{
\item{grouping_obj}{Grouping object made with \code{get_groups}}

\item{path}{Out directory for files}
}
\value{

}
\description{
Write grouping object to csv files.
}
