% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{xts_to_arima_model}
\alias{xts_to_arima_model}
\title{Model with ARIMA}
\usage{
xts_to_arima_model(xts_groupr, ..., is_auto_arima = TRUE,
  parallelize = FALSE, interval = "month")
}
\arguments{
\item{xts_groupr}{A \code{groupr} object produced with \code{extract_xts}.}

\item{...}{Arguments to be passed to \code{Arima} function (from the
\code{forecast} package).}

\item{is_auto_arima}{A boolean value to either specify a model explicitly or
to use the \code{auto.arima} function from the \code{forecast} package.}

\item{parallelize}{Whether or not to use the \code{parallel} package to do
calculations}

\item{interval}{Either "month", "week", or "day" for calculation.
Alternatively, you can provide a numeric value to be passed to
\code{frequency} for the \code{ts} object.}
}
\description{
This function allows the user to model all the combinations of \code{xts}
observations extracted using \code{extract_xts}. Because there can be many
columns in an \code{xts_groupr} object, there is the option to do model estimation in parallel.
}
