% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{extract_xts}
\alias{extract_xts}
\title{Extracts time series data from grouping object.}
\usage{
extract_xts(grouping_obj, groups, value_choice, date_col = "dd_dt")
}
\arguments{
\item{grouping_obj}{Grouping object created with \code{get_groups}}

\item{groups}{Groups that will be converted to time series data (\code{xts} object)}

\item{value_choice}{The column which will has values that will be in time series. Function applied will be \code{sum} but each value should be unique.}

\item{date_col}{The column with date data which will represent unique index for the returned \code{xts}}
}
\value{
An \code{xts} object with the date column cast against the groups column, using \code{sum} to summarize the value column
}
\description{
Extracts time series data from grouping object.
}
