% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_groups}
\alias{get_groups}
\title{Create grouping object.}
\usage{
get_groups(df, groups, functions = list(count = "n()"))
}
\arguments{
\item{df}{A data frame of raw data.}

\item{groups}{A character vector of column names in df that will be basis for aggregation.}

\item{functions}{A named list of functions for aggregation.}
}
\value{
A grouping object (list that aggregates functions by every potential combination)
}
\description{
This function takes a data frame as input and returns a grouping object as output. The grouping object is a list of lists of data frames, one data frame for every possible combination of groups.
For example, if there are two groups, A and B, four data frames are returned: data frames for A, B, both A and B, and a data frame summarizing the whole dataset.
}
\examples{
default_data <- data.frame(  my_group1 = c(rep("A", 4), rep("B", 4))
                           , my_group2 = c(rep("foo", 2), rep("bar", 2), rep("baz", 2), rep("potato", 2))
                           , my_data = runif(n = 8)
                          )
grouping_obj <- get_groups(df = default_data, groups = "my_group", functions = list(rando_sum = "sum(my_data)"))
}

