% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{infer_colgrps}
\alias{infer_colgrps}
\title{Set column grouping for an already structured dataframe}
\usage{
infer_colgrps(x, index_name = "group", sep = "_")
}
\description{
Set column grouping for an already structured dataframe
}
