% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_oriented.R
\name{print.groupr}
\alias{print.groupr}
\title{Printing groupr objects}
\usage{
\method{print}{groupr}(groupr_obj, include_colnames = F)
}
\arguments{
\item{groupr_obj}{A grouping object created with \code{get_groups}}

\item{include_colnames}{Boolean for whether or not to print colnames in tree}
}
\description{
Print a groupr object as tree.
}
\examples{
bball_groupr <- get_groups(baseball, groups = c("league", "team"), functions = list(toal_home_runs = "sum(HR, na.rm=T)", n="n()"))
print(bball_groupr)
print(bball_groupr, include_colnames = TRUE)
}
