% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{xts_to_arima_model}
\alias{xts_to_arima_model}
\title{Models time series list with an ARIMA model.}
\usage{
xts_to_arima_model(xts_gr_obj, ..., is_auto_arima)
}
\arguments{
\item{xts_gr_obj}{A list of \code{xts} data produced with \code{extract_xts}.}

\item{...}{Arguments to be passed to \code{Arima} function (from the \code{forecast} package).}

\item{is_auto_arima}{A boolean value to either specify a model explicitly or to use the \code{auto.arima} function from the \code{forecast} package.}
}
\description{
Models time series list with an ARIMA model.
}
